Trait any_key::AnyOrd

pub trait AnyOrd: Any {
    fn debug(&self, f: &mut Formatter) -> Result;
    fn eq(&self, other: &AnyOrd) -> bool;
    fn cmp(&self, other: &AnyOrd) -> Ordering;
}

Object-safe trait for dynamically typed totally ordered keys.

All Ord + 'static types automatically implement this trait.

On nightly, if the inner value implements Debug, then Debug for AnyOrd will use that implementation. Otherwise, it will only show a generic message.

Required Methods

fn debug(&self, f: &mut Formatter) -> Result

fn eq(&self, other: &AnyOrd) -> bool

fn cmp(&self, other: &AnyOrd) -> Ordering

Methods

impl AnyOrd

fn is<T: AnyOrd>(&self) -> bool

Returns true if the boxed type is the same as T

fn downcast_ref<T: AnyOrd>(&self) -> Option<&T>

Returns some reference to the boxed value if it is of type T, or None if it isn't.

unsafe fn downcast_ref_unchecked<T: AnyOrd>(&self) -> &T

Returns a reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

fn downcast_mut<T: AnyOrd>(&mut self) -> Option<&mut T>

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

unsafe fn downcast_mut_unchecked<T: AnyOrd>(&mut self) -> &mut T

Returns a mutable reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

impl AnyOrd

fn downcast<T: AnyOrd>(self: Box<Self>) -> Result<Box<T>, Box<Self>>

Returns the boxed value if it is of type T, or Err(Self) if it isn't.

unsafe fn downcast_unchecked<T: AnyOrd>(self: Box<Self>) -> Box<T>

Returns the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.